requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django
django-cors-headers==4.6.0
django-ninja==1.3.0
docutils==0.21.2
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pyzbar==0.1.9
qrcode==8.0
sqlparse==0.5.3
typing_extensions==4.12.2
